Variant dataset support for bars
Variant datasets are just lists of variant lists.
To use variant dataset, you need to provide mapping
between the chart elements and variant list indices.
E.g. in case of bar chart, one index maps for
row, another for column, and third for value.

The purpose of variant dataset is to make it possible
to have just one blob of data and parse different views
on it using different mappings.

Note that since variant datasets utilize variants,
they are not likely to be highly performant. However,
the impact is only felt once when the data is parsed
into format understood by datavis.

Change-Id: I0da39bd18528370afb5060368b0c3280e31e1cf7
Reviewed-by: Mika Salmela <mika.salmela@digia.com>