Optimize single item changes in bar/surface item models.
We are only able to optimize this in cases where rows and
columns of the model are directly mapped to rows and columns
of the data proxy. In other cases we do not know if the new
values of the changed data item in the model actually specify the
same row/column in our data proxy as the previous values.

Task-number: QTRD-2190
Change-Id: Ie014469ac894474900e5cfd6d91fd1a60353b1f7
Reviewed-by: Titta Heikkala <titta.heikkala@digia.com>