Fix build issue with Qt 5.2.1
QImage is not a QObject, and therefore QImage::Format doesn't seem
to be a supported type for a Q_PROPERTY, either. At least this was
the case in Qt 5.2.1. In Qt 5.3.1 this worked. In any case,
worked around the issue by removing textureFormat as a property.
This shouldn't be an issue, since texture cannot meaningfully be
manipulated from QML anyway.

Also fixed some compile warnings.

Change-Id: I257c1502d3340c49c24085fb8bf2e6176d875215
Reviewed-by: Tomi Korpipää <tomi.korpipaa@digia.com>