Simplify context handling at cleanup.
If the context exists at renderer deletion, it is possible that
it no longer has valid surface, which means it won't be possible
to set it back to current if we change context for deletion cleanup.
Since the current context will be one of our shared contexts anyway,
there is no need to do a dummy context for cleanup unless the current
context is null.

Change-Id: Ibabe081742beb975ee848ccb3690703ef5b027a9
Reviewed-by: Mika Salmela <mika.salmela@digia.com>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@digia.com>