Fix debug-and-release build
Datavis is part of the Qt release and there is no need to do magic hacks
to install the qmldir file.
This was supposed to help running examples separately, but we assume
that the lib is shipped with Qt proper.

The problem was that the qmldir file would be simultaneously copied into
the same location twice which will fail on Windows.

Change-Id: I1879fc5ff6303104392a66e54d5c278752a12655
Reviewed-by: Jędrzej Nowacki <jedrzej.nowacki@theqtcompany.com>
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@theqtcompany.com>
Reviewed-by: Tomi Korpipää <tomi.korpipaa@theqtcompany.com>