Do show module license on all help pages
Qt Data Visualization has a different main license than most other Qt
modules, which is easy to miss. We do show the license on the module
landing page and in the 'All modules' overview page, but if you
directly open an example or a class page you can still miss this.

This change adds the main license description to the header shared
by all pages of the module; we (mis)use the buildversion variable for
that, which was so far unused.

Right now qdoc does not 'linkify' the entry. It could add a generic
link to the landing page though, where the license is also described.

Task-number: QTBUG-70354
Change-Id: Ic4917541a6480d1cb3879f87e6aac0093dd1f3b3
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>