Update commercial license headers
Updated header.COMM to the files in tqtc-qtdatavis3d.
Examples, tests or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4160
Change-Id: Ib818c335f5d3c01f57e2fd240e06181d703de779
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>
Reviewed-by: Antti Kokko <antti.kokko@qt.io>