Fix direct rendering doesn't show anything
Rendering a graph directly to background didn't
work in Qt 6 since it uses RHI and there is
no way to not clear the color buffer
before rendering. The graph uses direct OpenGL
calls to render directly to background. Enabled
direct rendering to qml3doscilloscope example.
In addition, RenderDirectToBackground_NoClear got
deprecated.

When rendering directly to background,
using non-transparent qml item as a background
will hide the graph. This was already mentioned
in the documentation but not clearly enough.
Updated documentation for AbstractGraph3D.

Pick-to: 6.1
Fixes: QTBUG-90665
Change-Id: I53081bac382ab89573359886e4f5c4b41be8e86d
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>