Fix Q3DScatter memory leak
When using static optimization for scattergraph
and series mesh type is not point mesh,
the scatter3D renderer uses
ScatterObjectBufferHelper's fullLoad. In
the fullLoad old data buffers are deleted, if
m_meshDataLoaded is set to true. However, at the
beginning of the fullLoad, the flag was always
set to false so old data buffers were
never deleted. Moved the setting of
the m_meshDataLoaded to false after the deletion
of old buffers. Added similar code to
scatterpointbufferhelper for consistency.

Pick-to: 5.15 6.1
Fixes: QTBUG-80194
Change-Id: I2fd71af86b1fd73621074187f112fd9783df36df
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>