Fix: Nans at surfacedata row/column 0 render fail
If the first or last column or row of the data
given to surfacedataproxy contains NaN values,
the surfacedataproxy might fail to find the
limit values for the axis. Changed limiValues
function in surfacedataproxy so that if it
doesn't find min/max values from the first/last
column/row it continues from the next column/row.
Added a test to verify that the limit values are
found when first/last row contains NaN values.

Fixes: QTBUG-95112
Fixes: QTBUG-95109
Change-Id: I74a343e5bc8cae5b8020243859000fdb5ca7f86e
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>