Fix: Nans at surfacedata row/column 0 render fail
If the first or last column or row of the data
given to surfacedataproxy contains NaN values,
the surfacedataproxy might fail to find the
limit values for the axis. Changed limitValues
function in surfacedataproxy so that if it
doesn't find min/max values from the first/last
column/row it continues from the next column/row.
Added a test to verify that the limit values are
found when first/last row contains NaN values.

Pick-to: 6.2
Change-Id: Ica3eebb1c6072656f59394814c3fa5e334f12c54
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>