Update commercial license headers
Updated header.COMM to the files in tqtc-qtdatavis3d.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4940
Change-Id: I03e43b99e5a1adeebaa4e328c3126e6434cbc650
Reviewed-by: Tomi Korpipää <tomi.korpipaa@qt.io>