Doc: Use the \examplecategory macro to tag examples
The '\meta category' command was used for tagging examples with a
specific category, used in Qt Creators Welcome mode.

As we want to also generate lists of examples belonging to a category
elsewhere in the documentation, replace the command with a macro that
expands to the original \meta command and also adds the example to
a group using the \ingroup command. This way, the category names can
be used as arguments to the \generatelist or \annotatedlist commands.

Task-number: QTBUG-112731
Change-Id: Ibfd1179923b898259a5c96417aba8f7711e9c5ee
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 6c7984e6931fa7123c1faee8a8c91d159d293633)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>